code smell fixed
