+++ web/src/main/java/it/pagopa/selfcare/pagopa/backoffice/web/model/creditorinstituions/IbanRequestDto.java
@@ -1,4 +1,4 @@
-package it.pagopa.selfcare.pagopa.backoffice.web.model.creditorInstituions;
+package it.pagopa.selfcare.pagopa.backoffice.web.model.creditorinstituions;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
