+++ web/src/test/java/it/pagopa/selfcare/pagopa/backoffice/web/model/creditorinstituions/CreditorInstitutionDetailsResourceTest.java
@@ -1,4 +1,4 @@
-package it.pagopa.selfcare.pagopa.backoffice.web.model.creditorInstituions;
+package it.pagopa.selfcare.pagopa.backoffice.web.model.creditorinstituions;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,7 +19,7 @@ import static it.pagopa.selfcare.pagopa.TestUtils.mockInstance;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class CreditorInstitutionResourceTest {
+class CreditorInstitutionDetailsResourceTest {
     private Validator validator;
 
 
@@ -34,11 +34,16 @@ class CreditorInstitutionResourceTest {
         //given
         HashMap<String, Class<? extends Annotation>> toCheckMap = new HashMap<>();
         toCheckMap.put("enabled", NotNull.class);
+        toCheckMap.put("address", NotNull.class);
+        toCheckMap.put("pspPayment", NotNull.class);
+        toCheckMap.put("reportingFtp", NotNull.class);
+        toCheckMap.put("reportingZip", NotNull.class);
         toCheckMap.put("creditorInstitutionCode", NotBlank.class);
         toCheckMap.put("businessName", NotBlank.class);
 
 
-        CreditorInstitutionResource model = new CreditorInstitutionResource();
+
+        CreditorInstitutionDetailsResource model = new CreditorInstitutionDetailsResource();
         model.setCreditorInstitutionCode("");
         model.setBusinessName("");
         model.setEnabled(null);
@@ -58,7 +63,7 @@ class CreditorInstitutionResourceTest {
     @Test
     void validateNotNullFields() {
         //given
-        CreditorInstitutionResource creditorInstitutionResource = mockInstance(new CreditorInstitutionResource());
+        CreditorInstitutionDetailsResource creditorInstitutionResource = mockInstance(new CreditorInstitutionDetailsResource());
         //when
         Set<ConstraintViolation<Object>> violations = validator.validate(creditorInstitutionResource);
         //then
